+++ CHANGES.md
@@ -7,7 +7,7 @@
 - Add `Mesh.revolve`
 
 ## v0.1.2
-- Add v4 (4d vector) type at top level (use abstracted in Quaternion)
+- Add v4 (4d vector) type at top level (use abstracted in `Quaternion`)
 
 ## v0.1.1
 - Add `Mesh.drop_unused_points`